# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Evidence Locker API
  description: Capture immutable screenshots of web pages and retrieve them with a
    record id and key.
  version: 1.0.0
paths:
  /v1/evidence/capture:
    post:
      summary: Capture a URL and store an encrypted screenshot archive
      operationId: evidence_capture_v1_evidence_capture_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /v1/evidence/metadata:
    post:
      summary: Retrieve capture metadata (IP, timestamp, source URL) using an evidence
        code
      operationId: evidence_metadata_v1_evidence_metadata_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrieveRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceMetadataResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/evidence/retrieve:
    post:
      summary: 'Retrieve decrypted PNG screenshot using an evidence code (metadata:
        POST /v1/evidence/metadata)'
      operationId: evidence_retrieve_v1_evidence_retrieve_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrieveRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
            image/png: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /healthz:
    get:
      summary: Healthz
      operationId: healthz_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    CaptureRequest:
      properties:
        url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
          description: Public http(s) page to capture as evidence.
      type: object
      required:
      - url
      title: CaptureRequest
    CaptureResponse:
      properties:
        code:
          type: string
          title: Code
          description: Combined id and decryption key, formatted as XXXX-YYYYYYYY.
      type: object
      required:
      - code
      title: CaptureResponse
    ErrorBody:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorBody
    EvidenceMetadataResponse:
      properties:
        source_url:
          type: string
          title: Source Url
          description: URL that was captured.
        captured_at:
          type: string
          title: Captured At
          description: ISO 8601 UTC timestamp when the evidence was captured.
        client_ip:
          type: string
          title: Client Ip
          description: Client IP seen at capture time (from X-Forwarded-For or direct).
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
          description: User-Agent header at capture time, if any.
      type: object
      required:
      - source_url
      - captured_at
      - client_ip
      title: EvidenceMetadataResponse
      description: Metadata stored at capture time (client IP, UTC timestamp, source
        URL).
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    RetrieveRequest:
      properties:
        code:
          type: string
          title: Code
          description: Evidence code returned from capture (XXXX-YYYYYYYY).
      type: object
      required:
      - code
      title: RetrieveRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError